feat(generic): expose top level methods as JS APIs
ISSUES CLOSED: #54
