+++ src/electron-forge-init.js
@@ -1,19 +1,10 @@
-import debug from 'debug';
 import path from 'path';
 import program from 'commander';
 
-import initCustom from './init/init-custom';
-import initDirectory from './init/init-directory';
-import initGit from './init/init-git';
-import initNPM from './init/init-npm';
-import initStandardFix from './init/init-standard-fix';
-import initStarter from './init/init-starter-files';
-
 import './util/terminate';
+import { init } from './api';
 
-const d = debug('electron-forge:init');
-
-const main = async () => {
+(async () => {
   let dir = process.cwd();
   program
     .version(require('../package.json').version)
@@ -30,27 +21,12 @@ const main = async () => {
     })
     .parse(process.argv);
 
-  d(`Initializing in: ${dir}`);
+  const initOpts = {
+    dir,
+    interactive: true,
+    lintstyle: program.lintstyle,
+  };
+  if (program.template) initOpts.template = program.template;
 
-  if (!program.template) {
-    program.lintstyle = program.lintstyle.toLowerCase();
-    if (!['airbnb', 'standard'].includes(program.lintstyle)) {
-      d(`Unrecognized lintstyle argument: '${program.lintstyle}' -- defaulting to 'airbnb'`);
-      program.lintstyle = 'airbnb';
-    }
-  }
-
-  await initDirectory(dir);
-  await initGit(dir);
-  await initStarter(dir, program.template ? undefined : program.lintstyle);
-  await initNPM(dir, program.template ? undefined : program.lintstyle);
-  if (!program.template) {
-    if (program.lintstyle === 'standard') {
-      await initStandardFix(dir);
-    }
-  } else {
-    await initCustom(dir, program.template, program.lintstyle);
-  }
-};
-
-main();
+  await init(initOpts);
+})();
