+++ src/electron-forge-start.js
@@ -1,16 +1,11 @@
-import 'colors';
-import { spawn } from 'child_process';
 import fs from 'fs-promise';
 import path from 'path';
 import program from 'commander';
 
 import './util/terminate';
-import asyncOra from './util/ora-handler';
-import readPackageJSON from './util/read-package-json';
-import rebuild from './util/rebuild';
-import resolveDir from './util/resolve-dir';
+import { start } from './api';
 
-const main = async () => {
+(async () => {
   let dir = process.cwd();
   program
     .version(require('../package.json').version)
@@ -26,33 +21,10 @@ const main = async () => {
     })
     .parse(process.argv.slice(0, 2));
 
-  await asyncOra('Locating Application', async () => {
-    dir = await resolveDir(dir);
-    if (!dir) {
-      // eslint-disable-next-line no-throw-literal
-      throw 'Failed to locate startable Electron application';
-    }
+  await start({
+    dir,
+    interactive: true,
+    enableLogging: program.enableLogging,
+    args: process.argv.slice(2),
   });
-
-  const packageJSON = await readPackageJSON(dir);
-
-  await rebuild(dir, packageJSON.devDependencies['electron-prebuilt-compile'], process.platform, process.arch);
-
-  const spawnOpts = {
-    cwd: dir,
-    stdio: 'inherit',
-    env: Object.assign({}, process.env, program.enableLogging ? {
-      ELECTRON_ENABLE_LOGGING: true,
-      ELECTRON_ENABLE_STACK_DUMPING: true,
-    } : {}),
-  };
-  await asyncOra('Launching Application', async () => {
-    if (process.platform === 'win32') {
-      spawn(path.resolve(dir, 'node_modules/.bin/electron.cmd'), ['.'].concat(process.argv.slice(2)), spawnOpts);
-    } else {
-      spawn(path.resolve(dir, 'node_modules/.bin/electron'), ['.'].concat(process.argv.slice(2)), spawnOpts);
-    }
-  });
-};
-
-main();
+})();
